// src/shared/api/generated/models/scenarioActionDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ScenarioActionDtoConfig } from "./scenarioActionDtoConfig";

export interface ScenarioActionDto {
  config: ScenarioActionDtoConfig;
  /**
   * @nullable
   * @pattern ^[a-z][a-z0-9_-]{0,63}$
   */
  nextNodeKey?: string | null;
  /** @pattern ^[a-z][a-z0-9_-]{0,63}$ */
  nodeKey?: string;
  /** @minimum 0 */
  position: number;
  /** @pattern ^[A-Z][A-Z0-9_]*$ */
  type: string;
}